refactor(utils): 移除旧版 bilibili API
