--- src/utils/bilibili_api/_legacy/__init__.py
@@ -1,9 +1,0 @@
-"""
-@Author         : Ailitonia
-@Date           : 2024/10/25 10:53:24
-@FileName       : legacy
-@Project        : omega-miya
-@Description    : [Deactivated]Legacy bilibili API
-@GitHub         : https://github.com/Ailitonia
-@Software       : PyCharm 
-"""
