--- src/utils/bilibili_api/_legacy/model/__init__.py
@@ -1,40 +1,0 @@
-"""
-@Author         : Ailitonia
-@Date           : 2022/04/11 20:24
-@FileName       : __init__.py
-@Project        : nonebot2_miya 
-@Description    : Bilibili model
-@GitHub         : https://github.com/Ailitonia
-@Software       : PyCharm 
-"""
-
-from .dynamic import BilibiliDynamicCard, BilibiliDynamicModel, BilibiliUserDynamicModel
-from .interface import (
-    BilibiliWebConfirmRefreshInfo,
-    BilibiliWebCookieInfo,
-    BilibiliWebCookieRefreshInfo,
-    BilibiliWebInterfaceNav,
-    BilibiliWebInterfaceSpi,
-    BilibiliWebQrcodeGenerateInfo,
-    BilibiliWebQrcodePollInfo,
-)
-from .live_room import BilibiliLiveRoomModel, BilibiliUsersLiveRoomModel
-from .search import UserSearchingModel
-from .user import BilibiliUserModel
-
-__all__ = [
-    'BilibiliDynamicCard',
-    'BilibiliDynamicModel',
-    'BilibiliLiveRoomModel',
-    'BilibiliWebInterfaceNav',
-    'BilibiliWebInterfaceSpi',
-    'BilibiliWebCookieInfo',
-    'BilibiliWebQrcodeGenerateInfo',
-    'BilibiliWebQrcodePollInfo',
-    'BilibiliWebCookieRefreshInfo',
-    'BilibiliWebConfirmRefreshInfo',
-    'BilibiliUserDynamicModel',
-    'BilibiliUsersLiveRoomModel',
-    'BilibiliUserModel',
-    'UserSearchingModel'
-]
