--- src/utils/bilibili_api/future/__init__.py
@@ -1,23 +1,0 @@
-"""
-@Author         : Ailitonia
-@Date           : 2024/10/25 10:53:52
-@FileName       : future
-@Project        : omega-miya
-@Description    : bilibili API
-@GitHub         : https://github.com/Ailitonia
-@Software       : PyCharm 
-"""
-
-from .api import (
-    BilibiliCredential,
-    BilibiliDynamic,
-    BilibiliLive,
-    BilibiliUser,
-)
-
-__all__ = [
-    'BilibiliDynamic',
-    'BilibiliCredential',
-    'BilibiliLive',
-    'BilibiliUser',
-]
